kotlin/luhn/src/main/kotlin/Luhn.kt
object Luhn {
    fun isValid(input: String): Boolean {
        val sanitizedInput = input.replace(" ", "")

        return when {
            valid(sanitizedInput) -> checksum(sanitizedInput) % 10 == 0
            else -> false
        }
    }

    private fun valid(input: String) = input.all(Char::isDigit) && input.length > 1

    private fun checksum(input: String) = addends(input).sum()

    private fun addends(input: String) = input.digits().mapIndexed { i, j ->
        when {
            (input.length - i + 1) % 2 == 0 -> j
            j >= 5 -> j * 2 - 9
            else -> j * 2
        }
    }

    private fun String.digits() = this.map(Character::getNumericValue)
}